base/requirements.txt
torchvision
lightning[extra]
hydra-core
omegaconf
transformers
rich
pandas
pyarrow
tensorboard
scikit-learn
pyyaml